clean TE files
