+++ src/methods/infomeasures/infomeasures.jl
@@ -69,4 +69,4 @@ include("entropy_conditional/entropy_conditional.jl")
 include("entropy_joint.jl")
 include("mutualinfo/mutualinfo.jl")
 include("condmutualinfo/condmutualinfo.jl")
-include("transferentropy_new/transferentropy.jl")
+include("transferentropy/transferentropy.jl")
