--- src/methods/infomeasures/transferentropy/convenience/spike/spike.jl
@@ -1,3 +1,0 @@
-include("utils.jl")
-include("embed.jl")
-include("crossentropy.jl")
